requirements.txt
absl-py==1.3.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.6
commonmark==0.9.1
cycler==0.11.0
Cython==0.29.32
cytoolz==0.11.0
docker-pycreds==0.4.0
dpcpp-cpp-rt==2022.2.1
e2cnn==0.2.2
easydict==1.7
einops==0.6.0
filelock==3.8.2
fire==0.4.0
fonttools==4.34.4
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
huggingface-hub==0.11.1
hydra-core==1.2.0
idna==3.4
imageio==2.20.0
importlib-metadata==4.12.0
importlib-resources==5.10.0
intel-cmplr-lib-rt==2022.2.1
intel-cmplr-lic-rt==2022.2.1
intel-opencl-rt==2022.2.1
intel-openmp==2022.2.1
json-tricks==3.15.5
kiwisolver==1.4.4
lightly==1.2.35
lightly-utils==0.0.2
lightning-utilities==0.3.0
Mako==1.2.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl==2022.2.1
mkl-fft==1.3.1
mkl-service==2.4.0
mmcls==0.23.2
mmdet==2.25.1
model-index==0.1.11
mpmath==1.2.1
multidict==6.0.2
networkx==2.8.5
ninja==1.10.2.3
numpy==1.22.4
nvidia-pyindex==1.0.9
oauthlib==3.2.2
omegaconf==2.2.3
onnx==1.12.0
opencv-python==4.6.0.66
openmim==0.2.0
ordered-set==4.1.0
packaging==21.3
pandas==1.4.3
pathtools==0.1.2
Pillow==9.0.1
platformdirs==2.5.2
prettytable==3.3.0
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycosat==0.6.3
pycparser==2.21
pycuda==2022.1
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytools==2022.1.12
pytorch-lightning==1.8.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.0
sentry-sdk==1.9.8
setproctitle==1.3.2
Shapely==1.8.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
submitit==1.4.5
sympy==1.10.1
tabulate==0.8.10
tbb==2021.7.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
terminaltables==3.1.10
tifffile==2022.7.28
timm==0.6.7
tokenizers==0.13.2
toolz==0.12.0
torch==1.12.1
torch-summary==1.4.5
torchaudio==0.12.1
torchmetrics==0.10.3
torchvision==0.13.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.13.3
wcwidth==0.2.5
Werkzeug==2.2.2
yacs==0.1.8
yapf==0.32.0
yarl==1.8.1
zipp==3.8.1
